move metrics to a standalone package
