--- console/metrics.go
@@ -1,15 +1,0 @@
-package console
-
-import "expvar"
-
-var (
-	tps *expvar.Map
-)
-
-func init() {
-	initCounters()
-}
-
-func initCounters() {
-	tps = expvar.NewMap("tps")
-}
